+++ Project/SharedResources/SharedResources/Comms/Messages/TransactionRequestMessage.cs
@@ -4,7 +4,7 @@ namespace Shared.Comms.Messages
 {
     public class TransactionRequestMessage : Message
     {
-        public int Quantity { get; set; }//Positive value for buy, negative value for sell
+        public float Quantity { get; set; }//Positive value for buy, negative value for sell
         public ValuatedStock StockValue { get; set; }
         public int PortfolioId { get; set; }
 
@@ -19,7 +19,7 @@ namespace Shared.Comms.Messages
             StockValue = message.StockValue;
         }
 
-        public TransactionRequestMessage(int quantity, ValuatedStock stockValue)
+        public TransactionRequestMessage(float quantity, ValuatedStock stockValue)
         {
             Quantity = quantity;
             StockValue = stockValue;
